requirements.txt
tortoise-orm
tortoise-orm[asyncpg]
tortoise-orm[asyncmy]
tortoise-orm[psycopg]
tortoise-orm[asyncodbc]
tortoise-orm[accel]
mypy
isort
black
pylint